Refactoring and creating classes
